Move embed_url function and remove package per review
